refactor: fix golangci-lint issues across infrastructure packages
- Rename RabbitMQConsumer to Consumer, RabbitMQPublisher to Publisher, RedisPubSubPublisher to PubSubPublisher, RedisPubSubSubscriber to PubSubSubscriber (revive stutter)
- Unexport Client fields on redis publisher/subscriber
- Add doc comments to all exported types, vars, and functions
- Change Payload.UserID from int to int32 to eliminate gosec G115 overflow warning
- Fix defer pubsub.Close() errcheck in subscriber

+++ internal/infra/postgres/repository.go
@@ -1,3 +1,4 @@
+// Package postgres provides PostgreSQL-based implementations of domain interfaces.
 package postgres
 
 import (
@@ -6,23 +7,27 @@ import (
 	"go-realtime-chat/internal/infra/postgres/gen"
 )
 
+// MessageRepository implements the domain.MessageRepository interface using a PostgreSQL database.
 type MessageRepository struct {
 	q *gen.Queries
 }
 
+// NewMessageRepository creates a new MessageRepository backed by the given sqlc queries.
 func NewMessageRepository(q *gen.Queries) *MessageRepository {
 	return &MessageRepository{q: q}
 }
 
+// CreateNewMessage inserts a new chat message into the database.
 func (r *MessageRepository) CreateNewMessage(ctx context.Context, payload domain.Payload) error {
 	_, err := r.q.CreateMessage(ctx, gen.CreateMessageParams{
-		UserID: int32(payload.UserID),
+		UserID: payload.UserID,
 		ChatID: payload.ChatID,
 		Text:   payload.Message,
 	})
 	return err
 }
 
+// GetMessagesByChat retrieves all messages for a given chat ID.
 func (r *MessageRepository) GetMessagesByChat(ctx context.Context, chatID string) ([]domain.Payload, error) {
 	msgs, err := r.q.GetMessagesByChat(ctx, chatID)
 	if err != nil {
@@ -32,7 +37,7 @@ func (r *MessageRepository) GetMessagesByChat(ctx context.Context, chatID string
 	result := make([]domain.Payload, len(msgs))
 	for i, m := range msgs {
 		result[i] = domain.Payload{
-			UserID:  int(m.UserID),
+			UserID:  m.UserID,
 			ChatID:  m.ChatID,
 			Message: m.Text,
 		}
